// src/Services/API/Identity/API.Identity.Admin.Spa/Client/src/app/modules/clients/models/client.model.ts
export interface Client {
  id?: string;
  clientId: string;
  clientName: string;
  description?: string;
  clientUri?: string;
  logoUri?: string;
  enabled: boolean;
  protocolType: string;
  requireClientSecret: boolean;
  requirePkce: boolean;
  allowPlainTextPkce: boolean;
  requireConsent: boolean;
  allowRememberConsent: boolean;
  alwaysIncludeUserClaimsInIdToken: boolean;
  requireRequestObject: boolean;
  allowedGrantTypes?: string[];
  redirectUris?: string[];
  postLogoutRedirectUris?: string[];
  frontChannelLogoutUri?: string;
  frontChannelLogoutSessionRequired: boolean;
  backChannelLogoutUri?: string;
  backChannelLogoutSessionRequired: boolean;
  allowOfflineAccess: boolean;
  allowedScopes?: string[];
  allowedCorsOrigins?: string[];
  identityTokenLifetime: number;
  accessTokenLifetime: number;
  authorizationCodeLifetime: number;
  absoluteRefreshTokenLifetime: number;
  slidingRefreshTokenLifetime: number;
  consentLifetime?: number;
  refreshTokenUsage: number;
  updateAccessTokenClaimsOnRefresh: boolean;
  refreshTokenExpiration: number;
  accessTokenType: number;
  enableLocalLogin: boolean;
  identityProviderRestrictions?: string[];
  includeJwtId: boolean;
  alwaysSendClientClaims: boolean;
  clientClaimsPrefix?: string;
  pairWiseSubjectSalt?: string;
  userSsoLifetime?: number;
  userCodeType?: string;
  deviceCodeLifetime: number;
  allowAccessTokensViaBrowser: boolean;
  created: Date;
  updated?: Date;
  lastAccessed?: Date;
  nonEditable: boolean;
}